+++ demo/example.js
@@ -37,7 +37,15 @@ export const modal = `
 
   <Input type="text" name="subject" label="Subject" required />
   <Textarea name="message" label="Message" maxLength="500" />
-  <ConversationsSelect name="shareWith" label="Share with..." required />
+
+  <ConversationsSelect
+    name="shareWith"
+    label="Share with..."
+    required
+    include="public im"
+    excludeBotUsers
+    responseUrlEnabled
+  />
 
   <Input type="hidden" name="postId" value="xxxx" />
   <Input type="submit" value="Send" />