+++ src/block-kit/elements/Select.tsx
@@ -108,18 +108,41 @@ type UsersSelectProps = WithInputProps<
 >
 
 // Conversations select
-interface SingleConversationsSelectProps extends SingleSelectPropsBase {
+interface ConversationsSelectPropsBase {
+  /** [Experimental] Properties to filter conversations list are beta. */
+  include?: string | ('im' | 'mpim' | 'private' | 'public')[]
+
+  /** [Experimental] Properties to filter conversations list are beta. */
+  excludeExternalSharedChannels?: boolean
+
+  /** [Experimental] Properties to filter conversations list are beta. */
+  excludeBotUsers?: boolean
+}
+
+interface SingleConversationsSelectProps
+  extends SingleSelectPropsBase,
+    ConversationsSelectPropsBase {
   initialConversation?: string
   children?: undefined
 }
 
-interface MultiConversationsSelectProps extends MultiSelectPropsBase {
+interface SingleConversationsSelectInputProps {
+  responseUrlEnabled?: boolean
+}
+
+interface MultiConversationsSelectProps
+  extends MultiSelectPropsBase,
+    ConversationsSelectPropsBase {
   initialConversation?: string | string[]
   children?: undefined
 }
 
-type ConversationsSelectProps = WithInputProps<
-  SingleConversationsSelectProps | MultiConversationsSelectProps
+type ConversationsSelectProps = DistributedProps<
+  | WithInputProps<
+      SingleConversationsSelectProps,
+      SingleConversationsSelectInputProps
+    >
+  | WithInputProps<MultiConversationsSelectProps>
 >
 
 // Channels select
@@ -128,13 +151,18 @@ interface SingleChannelsSelectProps extends SingleSelectPropsBase {
   children?: undefined
 }
 
+interface SingleChannelsSelectInputProps {
+  responseUrlEnabled?: boolean
+}
+
 interface MultiChannelsSelectProps extends MultiSelectPropsBase {
   initialChannel?: string | string[]
   children?: undefined
 }
 
-type ChannelsSelectProps = WithInputProps<
-  SingleChannelsSelectProps | MultiChannelsSelectProps
+type ChannelsSelectProps = DistributedProps<
+  | WithInputProps<SingleChannelsSelectProps, SingleChannelsSelectInputProps>
+  | WithInputProps<MultiChannelsSelectProps>
 >
 
 // Options
@@ -161,7 +189,7 @@ type SelectFragmentObject<T extends 'options' | 'option_groups'> = Required<
   Pick<StaticSelect, T>
 >
 
-export const Option: JSXSlack.FC<OptionProps> = props => (
+export const Option: JSXSlack.FC<OptionProps> = (props) => (
   <ObjectOutput<OptionInternal>
     {...props}
     type="option"
@@ -169,7 +197,7 @@ export const Option: JSXSlack.FC<OptionProps> = props => (
   />
 )
 
-export const Optgroup: JSXSlack.FC<OptgroupProps> = props => (
+export const Optgroup: JSXSlack.FC<OptgroupProps> = (props) => (
   <ObjectOutput<OptgroupInternal> {...props} type="optgroup" />
 )
 
@@ -220,7 +248,7 @@ const generateFragments = (
   return fragment
 }
 
-export const SelectFragment: JSXSlack.FC<SelectFragmentProps> = props =>
+export const SelectFragment: JSXSlack.FC<SelectFragmentProps> = (props) =>
   makeSerializableToJSON(
     (() => {
       const opts = filter(props.children)
@@ -229,7 +257,7 @@ export const SelectFragment: JSXSlack.FC<SelectFragmentProps> = props =>
         return <ObjectOutput<SelectFragmentObject<'options'>> options={[]} />
 
       const { type } = opts[0].props
-      if (!opts.every(o => o.props.type === type))
+      if (!opts.every((o) => o.props.type === type))
         throw new Error(
           'Component for selection must only include either of <Option> and <Optgroup>.'
         )
@@ -238,7 +266,7 @@ export const SelectFragment: JSXSlack.FC<SelectFragmentProps> = props =>
         case 'option':
           return (
             <ObjectOutput<SelectFragmentObject<'options'>>
-              options={(opts as JSXSlack.Node<OptionInternal>[]).map(n =>
+              options={(opts as JSXSlack.Node<OptionInternal>[]).map((n) =>
                 createOption(n.props)
               )}
             />
@@ -247,9 +275,9 @@ export const SelectFragment: JSXSlack.FC<SelectFragmentProps> = props =>
           return (
             <ObjectOutput<SelectFragmentObject<'option_groups'>>
               option_groups={
-                (opts as JSXSlack.Node<OptgroupInternal>[]).map(n => ({
+                (opts as JSXSlack.Node<OptgroupInternal>[]).map((n) => ({
                   label: plainText(n.props.label),
-                  options: filter(n.props.children).map(o =>
+                  options: filter(n.props.children).map((o) =>
                     createOption(o.props)
                   ),
                 })) as any
@@ -262,17 +290,17 @@ export const SelectFragment: JSXSlack.FC<SelectFragmentProps> = props =>
     })()
   )
 
-export const Select: JSXSlack.FC<SelectProps> = props => {
+export const Select: JSXSlack.FC<SelectProps> = (props) => {
   const fragment = generateFragments(props.children)
 
   // Find initial option(s)
   const initialOptions: SlackOption[] = []
   const opts: SlackOption[] =
     fragment.options ||
-    flattenDeep(fragment.option_groups.map(og => og.options))
+    flattenDeep(fragment.option_groups.map((og) => og.options))
 
   for (const target of wrap(props.value || [])) {
-    const found = opts.find(o => o.value === target)
+    const found = opts.find((o) => o.value === target)
     if (found) initialOptions.push(found)
   }
 
@@ -296,10 +324,10 @@ export const Select: JSXSlack.FC<SelectProps> = props => {
   return props.label ? wrapInInput(element, props) : element
 }
 
-export const ExternalSelect: JSXSlack.FC<ExternalSelectProps> = props => {
+export const ExternalSelect: JSXSlack.FC<ExternalSelectProps> = (props) => {
   const minQueryLength = coerceToInteger(props.minQueryLength)
   const initialOptions = props.initialOption
-    ? wrap(props.initialOption).map(opt =>
+    ? wrap(props.initialOption).map((opt) =>
         isNode<OptionInternal>(opt) ? createOption(opt.props) : opt
       )
     : []
@@ -323,7 +351,7 @@ export const ExternalSelect: JSXSlack.FC<ExternalSelectProps> = props => {
   return props.label ? wrapInInput(element, props) : element
 }
 
-export const UsersSelect: JSXSlack.FC<UsersSelectProps> = props => {
+export const UsersSelect: JSXSlack.FC<UsersSelectProps> = (props) => {
   const element = props.multiple ? (
     <ObjectOutput<MultiUsersSelect>
       type="multi_users_select"
@@ -341,7 +369,30 @@ export const UsersSelect: JSXSlack.FC<UsersSelectProps> = props => {
   return props.label ? wrapInInput(element, props) : element
 }
 
-export const ConversationsSelect: JSXSlack.FC<ConversationsSelectProps> = props => {
+export const ConversationsSelect: JSXSlack.FC<ConversationsSelectProps> = (
+  props
+) => {
+  let filterComposition: SlackConversationsSelect['filter'] = {}
+  let { include } = props
+
+  if (include) {
+    if (!Array.isArray(include)) include = include.split(' ') as any[]
+
+    include = [...new Set(include)].filter((o) =>
+      ['im', 'mpim', 'private', 'public'].includes(o)
+    )
+
+    if (include.length > 0) filterComposition.include = include
+  }
+
+  if (props.excludeBotUsers !== undefined)
+    filterComposition.exclude_bot_users = !!props.excludeBotUsers
+
+  if (props.excludeExternalSharedChannels !== undefined)
+    filterComposition.exclude_external_shared_channels = !!props.excludeExternalSharedChannels
+
+  if (Object.keys(filterComposition).length === 0) filterComposition = undefined
+
   const element = props.multiple ? (
     <ObjectOutput<MultiConversationsSelect>
       type="multi_conversations_select"
@@ -349,19 +400,26 @@ export const ConversationsSelect: JSXSlack.FC<ConversationsSelectProps> = props 
       initial_conversations={
         props.initialConversation ? wrap(props.initialConversation) : undefined
       }
+      filter={filterComposition}
     />
   ) : (
     <ObjectOutput<SlackConversationsSelect>
       type="conversations_select"
       {...baseProps(props)}
       initial_conversation={props.initialConversation}
+      filter={filterComposition}
+      response_url_enabled={
+        props.responseUrlEnabled !== undefined
+          ? !!props.responseUrlEnabled
+          : undefined
+      }
     />
   )
 
   return props.label ? wrapInInput(element, props) : element
 }
 
-export const ChannelsSelect: JSXSlack.FC<ChannelsSelectProps> = props => {
+export const ChannelsSelect: JSXSlack.FC<ChannelsSelectProps> = (props) => {
   const element = props.multiple ? (
     <ObjectOutput<MultiChannelsSelect>
       type="multi_channels_select"
@@ -371,10 +429,15 @@ export const ChannelsSelect: JSXSlack.FC<ChannelsSelectProps> = props => {
       }
     />
   ) : (
-    <ObjectOutput<SlackChannelsSelect>
+    <ObjectOutput<SlackChannelsSelect & { response_url_enabled?: boolean }>
       type="channels_select"
       {...baseProps(props)}
       initial_channel={props.initialChannel}
+      response_url_enabled={
+        props.responseUrlEnabled !== undefined
+          ? !!props.responseUrlEnabled
+          : undefined
+      }
     />
   )