+++ src/legacy/jsx.ts
@@ -14,10 +14,11 @@ const legacyJsxToHtml = (
   const text = () => children.join('')
 
   const isChild = (...elements: string[]) =>
-    parents.length > 0 && elements.some(e => e === parents[parents.length - 1])
+    parents.length > 0 &&
+    elements.some((e) => e === parents[parents.length - 1])
 
   const isDescendant = (...elements: string[]) =>
-    elements.some(e => parents.includes(e))
+    elements.some((e) => parents.includes(e))
 
   if (name === 'br') return '<br />'
   if (isChild('pre', 'code') && !['a', 'time'].includes(name)) return text()